fdsfsdf

+++ trab2ed/freq_table.h
@@ -10,7 +10,7 @@ Freq_Table* init_freq_table(void);
 
 void insert_freq_table(Freq_Table* f_tbl, int i );
 
-int get_freq_table(Freq_Table* f_tbl, int i);
+long long unsigned int get_freq_table(Freq_Table* f_tbl, int i);
 
 void free_freq_table(Freq_Table* f_tble);
